refactor: move submit form to separate component
